Refactor View Data single-location summary into new component
and add `FcViewDataAggregate` as a placeholder for aggregate view.
