+++ web/components/FcDrawerViewData.vue
@@ -20,12 +20,29 @@
         </h1>
         <div class="label mt-5">
           <span v-if="locationFeatureType !== null">
-            {{locationFeatureType.description}} &#x2022;
+            {{locationFeatureType.description}}
           </span>
-          <span>{{studySummaryHeaderText}}</span>
+          <v-progress-circular
+            v-if="loading"
+            color="primary"
+            indeterminate
+            :size="20"
+            :width="2" />
+          <span v-else>
+            &#x2022; {{studySummaryHeaderText}}
+          </span>
         </div>
         <div class="d-flex mt-4">
-          <FcSummaryPoi :poi-summary="poiSummary" />
+          <v-progress-circular
+            v-if="loading"
+            class="ma-3"
+            color="primary"
+            indeterminate
+            :size="20"
+            :width="2" />
+          <FcSummaryPoi
+            v-else
+            :location="location" />
           <v-spacer></v-spacer>
           <FcButton
             type="secondary"
@@ -37,138 +54,17 @@
       </div>
     </header>
     <section class="flex-grow-1 flex-shrink-1 overflow-y-auto">
+      <v-divider></v-divider>
       <v-progress-linear
         v-if="loading"
         indeterminate />
       <template v-else>
-        <v-divider></v-divider>
-
-        <section>
-          <header class="pa-5">
-            <div class="align-center d-flex">
-              <h2 class="headline">Collisions</h2>
-              <div class="pl-3 subtitle-1">{{collisionTotal}} total</div>
-              <v-spacer></v-spacer>
-              <FcDialogCollisionFilters
-                v-if="showFiltersCollision"
-                v-model="showFiltersCollision"
-                v-bind="filtersCollision"
-                @set-filters="setFiltersCollision">
-              </FcDialogCollisionFilters>
-              <FcButton
-                v-if="collisionTotal > 0"
-                type="secondary"
-                @click.stop="showFiltersCollision = true">
-                <v-icon
-                  :color="colorIconFilterCollision"
-                  left>mdi-filter-variant</v-icon>
-                Filter
-              </FcButton>
-            </div>
-
-            <div
-              v-if="filterChipsCollision.length > 0"
-              class="mt-5">
-              <v-chip
-                v-for="(filterChip, i) in filterChipsCollision"
-                :key="i"
-                class="mb-2 mr-2 primary--text"
-                color="light-blue lighten-5"
-                @click="removeFilterCollision(filterChip)">
-                <v-icon left>mdi-check</v-icon>
-                {{filterChip.label}}
-              </v-chip>
-            </div>
-          </header>
-
-          <FcDataTableCollisions
-            :collision-summary="collisionSummary"
-            :loading="loadingCollisions"
-            @show-reports="actionShowReportsCollision" />
-        </section>
-
-        <v-divider></v-divider>
-
-        <section>
-          <header class="pa-5">
-            <div class="align-center d-flex">
-              <h2 class="headline">Studies</h2>
-              <div class="pl-3 subtitle-1">{{studyTotal}} total</div>
-              <v-spacer></v-spacer>
-              <FcDialogStudyFilters
-                v-if="showFiltersStudy"
-                v-model="showFiltersStudy"
-                v-bind="filtersStudy"
-                @set-filters="setFiltersStudy">
-              </FcDialogStudyFilters>
-              <FcButton
-                v-if="studyTotal > 0"
-                type="secondary"
-                @click.stop="showFiltersStudy = true">
-                <v-icon
-                  :color="colorIconFilterStudy"
-                  left>mdi-filter-variant</v-icon>
-                Filter
-              </FcButton>
-              <FcButton
-                class="ml-3"
-                type="primary"
-                @click="actionRequestStudy">
-                <v-icon left>mdi-plus-box</v-icon>
-                Request Study
-              </FcButton>
-            </div>
-
-            <div
-              v-if="filterChipsStudy.length > 0"
-              class="mt-5">
-              <v-chip
-                v-for="(filterChip, i) in filterChipsStudy"
-                :key="i"
-                class="mb-2 mr-2 primary--text"
-                color="light-blue lighten-5"
-                @click="removeFilterStudy(filterChip)">
-                <v-icon left>mdi-check</v-icon>
-                {{filterChip.label}}
-              </v-chip>
-            </div>
-          </header>
-          <div
-            v-if="studyTotal === 0"
-            class="my-8 py-12 secondary--text text-center">
-            There are no studies for this location,<br>
-            please request a study if necessary
-          </div>
-          <div
-            v-else-if="studySummary.length === 0"
-            class="my-8 py-12 secondary--text text-center">
-            No studies match the active filters,<br>
-            clear one or more filters to see studies
-          </div>
-          <FcDataTableStudies
-            v-else
-            :loading="loadingStudies"
-            :study-summary="studySummary"
-            @show-reports="actionShowReportsStudy" />
-          <div class="pa-5">
-            <div
-              v-for="studyRequest in studyRequestsPending"
-              :key="studyRequest.id"
-              class="align-center d-flex">
-              <v-icon
-                color="warning"
-                left>mdi-information</v-icon>
-              <div>
-                {{studyRequest.studyType.label}}
-                has been requested on {{studyRequest.createdAt | date}}.
-                <router-link :to="{
-                  name: 'requestStudyView',
-                  params: { id: studyRequest.id },
-                }">View details.</router-link>
-              </div>
-            </div>
-          </div>
-        </section>
+        <FcViewDataDetail
+          v-if="detailView"
+          :location="location" />
+        <FcViewDataAggregate
+          v-else
+          :locations="locations" />
       </template>
     </section>
   </div>
@@ -183,81 +79,46 @@ import {
 } from 'vuex';
 
 import {
-  AuthScope,
   LocationMode,
   LocationSelectionType,
 } from '@/lib/Constants';
-import {
-  getCollisionsByCentrelineSummary,
-  getCollisionsByCentrelineTotal,
-  getPoiByCentrelineSummary,
-  getStudiesByCentrelineSummary,
-  getStudiesByCentrelineTotal,
-  getStudyRequestsByCentrelinePending,
-} from '@/lib/api/WebApi';
+import { getStudiesByCentrelineSummary } from '@/lib/api/WebApi';
 import CompositeId from '@/lib/io/CompositeId';
 import DateTime from '@/lib/time/DateTime';
 import TimeFormatters from '@/lib/time/TimeFormatters';
-import FcDataTableCollisions from '@/web/components/FcDataTableCollisions.vue';
-import FcDataTableStudies from '@/web/components/FcDataTableStudies.vue';
-import FcDialogCollisionFilters from '@/web/components/dialogs/FcDialogCollisionFilters.vue';
-import FcDialogStudyFilters from '@/web/components/dialogs/FcDialogStudyFilters.vue';
+import FcViewDataAggregate from '@/web/components/data/FcViewDataAggregate.vue';
+import FcViewDataDetail from '@/web/components/data/FcViewDataDetail.vue';
 import FcButton from '@/web/components/inputs/FcButton.vue';
 import FcSelectorSingleLocation from '@/web/components/inputs/FcSelectorSingleLocation.vue';
 import FcSelectorMultiLocation from '@/web/components/inputs/FcSelectorMultiLocation.vue';
 import FcSummaryPoi from '@/web/components/location/FcSummaryPoi.vue';
-import FcMixinAuthScope from '@/web/mixins/FcMixinAuthScope';
 import FcMixinRouteAsync from '@/web/mixins/FcMixinRouteAsync';
 
 export default {
   name: 'FcDrawerViewData',
   mixins: [
-    FcMixinAuthScope,
     FcMixinRouteAsync,
   ],
   components: {
     FcButton,
-    FcDataTableCollisions,
-    FcDataTableStudies,
-    FcDialogCollisionFilters,
-    FcDialogStudyFilters,
     FcSelectorMultiLocation,
     FcSelectorSingleLocation,
     FcSummaryPoi,
+    FcViewDataAggregate,
+    FcViewDataDetail,
   },
   data() {
     return {
-      collisionSummary: {
-        amount: 0,
-        ksi: 0,
-        validated: 0,
-      },
-      collisionTotal: 0,
       detailView: false,
-      loadingCollisions: false,
-      loadingStudies: false,
-      poiSummary: {
-        school: null,
-      },
-      showFiltersCollision: false,
-      showFiltersStudy: false,
-      studyRequestsPending: [],
       studySummary: [],
-      studyTotal: 0,
     };
   },
   computed: {
-    colorIconFilterCollision() {
-      if (this.filterChipsCollision.length === 0) {
-        return 'unselected';
+    location() {
+      if (this.locations.length === 0) {
+        return null;
       }
-      return 'primary';
-    },
-    colorIconFilterStudy() {
-      if (this.filterChipsStudy.length === 0) {
-        return 'unselected';
-      }
-      return 'primary';
+      return this.locations[0];
     },
     studySummaryHeaderText() {
       const n = this.studySummary.length;
@@ -271,10 +132,6 @@ export default {
       const mostRecentDateStr = TimeFormatters.formatDefault(mostRecentDate);
       return `${nStr} (${mostRecentDateStr})`;
     },
-    hasPoisNearby() {
-      const { hospital, school } = this.poiSummary;
-      return hospital !== null || school !== null;
-    },
     ...mapState('viewData', ['filtersCollision', 'filtersStudy']),
     ...mapState([
       'auth',
@@ -282,12 +139,6 @@ export default {
       'locationMode',
       'locations',
       'locationsSelection',
-    ]),
-    ...mapGetters('viewData', [
-      'filterChipsCollision',
-      'filterChipsStudy',
-      'filterParamsCollision',
-      'filterParamsStudy',
     ]),
     ...mapGetters([
       'locationFeatureType',
@@ -297,24 +148,6 @@ export default {
     ]),
   },
   watch: {
-    async filterParamsCollision() {
-      this.loadingCollisions = true;
-      const collisionSummary = await getCollisionsByCentrelineSummary(
-        this.locations,
-        this.filterParamsCollision,
-      );
-      this.collisionSummary = collisionSummary;
-      this.loadingCollisions = false;
-    },
-    async filterParamsStudy() {
-      this.loadingStudies = true;
-      const studySummary = await getStudiesByCentrelineSummary(
-        this.locations,
-        this.filterParamsStudy,
-      );
-      this.studySummary = studySummary;
-      this.loadingStudies = false;
-    },
     locationMode() {
       this.detailView = false;
     },
@@ -357,56 +190,14 @@ export default {
     actionAddLocation() {
       this.setLocationMode(LocationMode.MULTI_EDIT);
     },
-    actionRequestStudy() {
-      this.$router.push({ name: 'requestStudyNew' });
-    },
-    actionShowReportsCollision() {
-      const params = this.locationsRouteParams;
-      this.$router.push({
-        name: 'viewCollisionReportsAtLocation',
-        params,
-      });
-    },
-    actionShowReportsStudy({ category: { studyType } }) {
-      const params = this.locationsRouteParams;
-      this.$router.push({
-        name: 'viewStudyReportsAtLocation',
-        params: {
-          ...params,
-          studyTypeName: studyType.name,
-        },
-      });
-    },
     async loadAsyncForRoute(to) {
       const { s1, selectionTypeName } = to.params;
       const features = CompositeId.decode(s1);
       const selectionType = LocationSelectionType.enumValueOf(selectionTypeName);
       await this.initLocations({ features, selectionType });
 
-      const tasks = [
-        getCollisionsByCentrelineSummary(this.locations, this.filterParamsCollision),
-        getCollisionsByCentrelineTotal(this.locations),
-        getPoiByCentrelineSummary(this.locations[0]),
-        getStudiesByCentrelineSummary(this.locations, this.filterParamsStudy),
-        getStudiesByCentrelineTotal(this.locations),
-      ];
-      if (this.hasAuthScope(AuthScope.STUDY_REQUESTS)) {
-        tasks.push(getStudyRequestsByCentrelinePending(this.locations));
-      }
-      const [
-        collisionSummary,
-        collisionTotal,
-        poiSummary,
-        studySummary,
-        studyTotal,
-        studyRequestsPending = [],
-      ] = await Promise.all(tasks);
-      this.collisionSummary = collisionSummary;
-      this.collisionTotal = collisionTotal;
-      this.poiSummary = poiSummary;
-      this.studyRequestsPending = studyRequestsPending;
+      const studySummary = await getStudiesByCentrelineSummary(this.locations, {});
       this.studySummary = studySummary;
-      this.studyTotal = studyTotal;
     },
     ...mapMutations(['setLocationMode']),
     ...mapMutations('viewData', [
